pyproject.toml
[tool.poetry]
name = "kafka-materializer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
aiokafka = "^0.7.2"
typer = "^0.4.0"
cassandra-driver = "^3.25.0"
kafka-python = "^2.0.2"
pytest-asyncio = "^0.16.0"
orjson = "^3.6.5"
python-dotenv = "^0.19.2"
Faker = "^10.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
flake8 = "^4.0.1"
mypy = "^0.920"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
materializer = 'kafka_materializer.commands:app'